services/autodoc-thinker/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.10.0
pypdf==4.0.1
python-docx==1.1.0
redis==5.0.1
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
